requirements.txt
Flask>=2.2.2,<3
Flask-SQLAlchemy>= 2.5.1
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.5
Flask-Cors==3.0.9
requests==2.25.0
mysqlclient==2.0.1
pika==1.1.0
python-dotenv==1.0.0